A special heroku-specific environment
As fisheye isn't meant to only run on Heroku, but actually be self-hosted on
your own hardware, 'production' can be a variable number of things. However, it
is also a valid target, and useful for testing the implementations (hard to
verify an OpenID running on 'localhost', for instance) amongst various websites.

This way, production is a relatively clean example while Heroku can be all
'dirty' without getting in the user's way.